+++ CHANGELOG.md
@@ -4,6 +4,11 @@ This file documents notable changes between versions of quivr.
 
 ## Unreleased
 
+### Fixed
+
+- Loading Table data from pyarrow ChunkedArrays (like with
+  from_kwargs) is now less buggy and faster (#59)
+  
 ### Removed
 
 - The `from_data`, `from_list`, `from_rows`, and `from_pydict`